+++ repos/cache/src/commonMain/kotlin/dev/inmo/micro_utils/repos/cache/cache/KVCache.kt
@@ -1,0 +1,6 @@
+package dev.inmo.micro_utils.repos.cache.cache
+
+import dev.inmo.micro_utils.repos.*
+
+interface KVCache<K, V> : KeyValueRepo<K, V>
+